Grande remap do projeto (Todos os creates finalizados)

+++ src/main/java/com/SCX/ControleDeExame/service/PatientService.java
@@ -5,9 +5,11 @@ import com.SCX.ControleDeExame.dataTransferObject.patientDTO.GetAllPatientDTO;
 import com.SCX.ControleDeExame.dataTransferObject.patientDTO.PatientDTO;
 import com.SCX.ControleDeExame.domain.auth.Auth;
 import com.SCX.ControleDeExame.domain.patient.Patient;
-import com.SCX.ControleDeExame.domain.roleEnum.RoleEnum;
+import com.SCX.ControleDeExame.domain.role.Role;
+import com.SCX.ControleDeExame.domain.role.RoleEnum;
 import com.SCX.ControleDeExame.repository.AuthRepository;
 import com.SCX.ControleDeExame.repository.PatientRepository;
+import com.SCX.ControleDeExame.repository.RoleRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -26,37 +28,47 @@ public class PatientService {
     PatientRepository patientRepository;
     @Autowired
     AuthRepository authRepository;
-
+    @Autowired
+    RoleRepository roleRepository;
 
 
     public void registerPatient(PatientDTO data) {
 
-        String senhaTemp = UUID.randomUUID().toString().substring(0,8);
-        String token = UUID.randomUUID().toString();
-        Boolean status = false;
-        Timestamp expirationToken = Timestamp.valueOf(LocalDateTime.now().plusDays(1));
-        Boolean token_status = true;
+        Role patient = roleRepository.findByName("Patient");
+        Auth user = (Auth) authRepository.findByUsernameKey(data.email());
 
-        String encryptedPassword = new BCryptPasswordEncoder().encode(senhaTemp);
-        Auth newAuth = new Auth(data.cpf(), encryptedPassword, RoleEnum.PATIENT, token, status, expirationToken, token_status);
-        authRepository.save(newAuth);
+        if (user.getUsernameKey() != null) {
 
+            user.getRoles().add(patient);
+            authRepository.save(user);
+        } else {
 
-        try {
-            Patient newPatient = new Patient();
-            newPatient.setCpf(data.cpf());
-            newPatient.setName(data.name());
-            newPatient.setEmail(data.email());
-            newPatient.setAuth_id(newAuth);
-            patientRepository.save(newPatient);
+            String senhaTemp = UUID.randomUUID().toString().substring(0, 8);
+            String token = UUID.randomUUID().toString();
+            Boolean status = false;
+            Timestamp expirationToken = Timestamp.valueOf(LocalDateTime.now().plusDays(1));
+            Boolean token_status = true;
 
-        } catch (Exception e) {
-            authRepository.delete(newAuth);
-            e.printStackTrace();
-            throw e;
+            String encryptedPassword = new BCryptPasswordEncoder().encode(senhaTemp);
+            Auth newAuth = new Auth(data.email(), encryptedPassword, token, status, expirationToken, token_status);
+
+            newAuth.getRoles().add(patient);
+            authRepository.save(newAuth);
+
+            try {
+                Patient newPatient = new Patient();
+                newPatient.setCpf(data.cpf());
+                newPatient.setName(data.name());
+                newPatient.setEmail(data.email());
+                newPatient.setAuthId(newAuth);
+                patientRepository.save(newPatient);
+
+            } catch (Exception e) {
+                authRepository.delete(newAuth);
+                e.printStackTrace();
+                throw e;
+            }
         }
-
-
 
 
     }
@@ -68,7 +80,7 @@ public class PatientService {
 
     }
 
-    public Patient updatePatient (PatientDTO data, UUID uuid){
+    public Patient updatePatient(PatientDTO data, UUID uuid) {
         Patient patientUpdate = patientRepository.findById(uuid).orElseThrow(() -> new EntityNotFoundException("paciente não encontrado"));
 
         patientUpdate.setTelephone(data.telephone());
@@ -78,17 +90,17 @@ public class PatientService {
 
     }
 
-    public List<GetAllPatientDTO> getAllPatient (){
+    public List<GetAllPatientDTO> getAllPatient() {
 
         return patientRepository.findAll().stream().map(GetAllPatientDTO::new).toList();
     }
 
-    public Patient getPatientByEmail(PatientDTO data){
+    public Patient getPatientByEmail(PatientDTO data) {
 
         return patientRepository.findByEmail(data.email());
     }
 
-    public Patient getPatientById(RequestTokenDTO data){
+    public Patient getPatientById(RequestTokenDTO data) {
         return patientRepository.findById(UUID.fromString(data.Token())).orElseThrow(() -> new EntityNotFoundException("paciente não encontrado"));
     }
 }
